se finalizo con la implementacion de roles y permisos ocultando botones y protegiendo rutas deacuerdo al rol tambien se corrigio el formulario de editar personas con la opcion de añadir o eliminar el rol de lider de dosimetria para mas de una sede, falta implementar eso para los demas roles. Tambien se corrigio el modulo de novedades

+++ app/Http/Controllers/PersonasrolesController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Persona;
+use App\Models\Personasedes;
 use App\Models\Personasroles;
 use App\Models\User;
 use Illuminate\Http\Request;
@@ -15,15 +16,26 @@ class PersonasrolesController extends Controller
         $this->middleware('auth');
     }
      //
-    public function destroy(Personasroles $personarol){
+    public function destroy(Personasroles $personarol, $sede){
+        
         if($personarol->role_id == 3){
-            $user = User::where('persona_id', '=', $personarol->persona_id)->get();
-            $usuario = User::find($user[0]->id);
-            $usuario->delete();
-            $personasede = Persona::where('id_persona', '=', $personarol->persona_id)
-            ->update(['lider_dosimetria' => NULL ]);
+            $personasede = Personasedes::where('persona_id', '=', $personarol->persona_id)
+            ->where('lider_dosimetria', '=', TRUE)->get();
+            if(count($personasede) == 1){
+                $user = User::where('persona_id', '=', $personarol->persona_id)->get();
+                $usuario = User::find($user[0]->id);
+                $usuario->delete();
+                $personasede = Personasedes::where('persona_id', '=', $personarol->persona_id)
+                ->update(['lider_dosimetria' => NULL]);
+                $personarol->delete();
+            }else{
+                $personasede = Personasedes::where('persona_id', '=', $personarol->persona_id)
+                ->where('sede_id', '=', $sede)
+                ->update(['lider_dosimetria' => NULL]);
+            }
+        }else{
+            $personarol->delete();
         }
-        $personarol->delete();
 
         /* return $personarol; */
         return back()->with('eliminado', 'ok');
